requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
thefuzz==0.22.1
python-Levenshtein==0.26.1
